+++ src/Symfony/Component/Form/ChoiceList/Loader/IntlCallbackChoiceLoader.php
@@ -24,13 +24,7 @@ class IntlCallbackChoiceLoader extends CallbackChoiceLoader
      */
     public function loadChoicesForValues(array $values, $value = null)
     {
-        // Optimize
-        $values = array_filter($values);
-        if (empty($values)) {
-            return [];
-        }
-
-        return $this->loadChoiceList($value)->getChoicesForValues($values);
+        return parent::loadChoicesForValues(array_filter($values), $value);
     }
 
     /**
@@ -40,15 +34,12 @@ class IntlCallbackChoiceLoader extends CallbackChoiceLoader
     {
         // Optimize
         $choices = array_filter($choices);
-        if (empty($choices)) {
-            return [];
-        }
 
         // If no callable is set, choices are the same as values
         if (null === $value) {
             return $choices;
         }
 
-        return $this->loadChoiceList($value)->getValuesForChoices($choices);
+        return parent::loadValuesForChoices($choices, $value);
     }
 }