+++ src/Symfony/Bridge/Doctrine/CHANGELOG.md
@@ -7,6 +7,7 @@ CHANGELOG
  * changed guessing of DECIMAL to set the `input` option of `NumberType` to string
  * deprecated not passing an `IdReader` to the `DoctrineChoiceLoader` when query can be optimized with a single id field
  * deprecated passing an `IdReader` to the `DoctrineChoiceLoader` when entities have a composite id
+ * added an `AbstractChoiceLoader` to simplify implementations and handle global optimizations
 
 4.2.0
 -----